fix(tests): move filter_openai_messages_for_api out of wonderwall package
test_unit_social_agent_messages.py imported it from
wonderwall.social_agent.agent, which pulls in wonderwall/__init__.py's
numpy/camel-ai import chain — not installed in the thin unit-test CI job,
breaking collection for the whole suite.

Move the dependency-free filter into app/utils/llm_message_filter.py and
have agent.py import it from there instead.
